pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xauusd-ichimoku-agent"
version = "0.1.0"
description = "XAU/USD Ichimoku trading agent for The5ers prop firm"
requires-python = ">=3.11"
dependencies = [
    # Config & validation
    "pydantic>=2.0",
    "pyyaml>=6.0",

    # Data processing
    "numpy>=1.26",
    "pandas>=2.1",
    "requests>=2.31",

    # Backtesting & optimisation
    "vectorbt>=0.26",
    "optuna>=3.4",

    # Database
    "psycopg2-binary>=2.9",
    "pgvector>=0.2",

    # Machine learning & statistics
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "hmmlearn>=0.3",

    # Visualisation
    "matplotlib>=3.8",
    "plotly>=5.18",
]

[project.optional-dependencies]
# MetaTrader5 is Windows-only; install separately on execution host
mt5 = ["MetaTrader5>=5.0"]

dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
